Add subtitle kind
