+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add `Package` kind
 - Add `Playlist` kind
 - Add `Rom` kind
+- Add `Subtitle` kind
 - Add `reader-ebml` feature
 - Change the return type of `FileFormat::short_name` from `&str` to `Option<&str>`
 - Drop `reader-mkv` and `accuracy-mkv` features
