+++ src/lib.rs
@@ -305,6 +305,8 @@ pub enum Kind {
     /// Plain text, source codes, markup languages, and other types of files that contain written
     /// text.
     Text,
+    /// Subtitles and captions.
+    Subtitle,
     /// Movies, animations, and other types of files that contain moving images, possibly with color
     /// and coordinated sound.
     Video,
