+++ tests/application.rs
@@ -283,41 +283,9 @@ fn test_sqlite3() {
 }
 
 #[test]
-fn test_subrip_text() {
-    let fmt = FileFormat::from_file("fixtures/application/sample.srt").unwrap();
-    assert_eq!(fmt, FileFormat::SubripText);
-}
-
-#[test]
 fn test_tasty() {
     let fmt = FileFormat::from_file("fixtures/application/sample.tasty").unwrap();
     assert_eq!(fmt, FileFormat::Tasty);
-}
-
-#[test]
-fn test_timed_text_markup_language_1() {
-    let fmt = FileFormat::from_file("fixtures/application/sample1.ttml").unwrap();
-    assert_eq!(fmt, FileFormat::TimedTextMarkupLanguage);
-}
-
-#[cfg(feature = "reader-xml")]
-#[test]
-fn test_timed_text_markup_language_2() {
-    let fmt = FileFormat::from_file("fixtures/application/sample2.ttml").unwrap();
-    assert_eq!(fmt, FileFormat::TimedTextMarkupLanguage);
-}
-
-#[test]
-fn test_universal_subtitle_format_1() {
-    let fmt = FileFormat::from_file("fixtures/application/sample1.usf").unwrap();
-    assert_eq!(fmt, FileFormat::UniversalSubtitleFormat);
-}
-
-#[cfg(feature = "reader-xml")]
-#[test]
-fn test_universal_subtitle_format_2() {
-    let fmt = FileFormat::from_file("fixtures/application/sample2.usf").unwrap();
-    assert_eq!(fmt, FileFormat::UniversalSubtitleFormat);
 }
 
 #[test]
